Refactor build system and improve cross-platform support
- Remove deprecated CMake and batch build scripts
- Update native wrapper to properly handle parallel arcs
- Fix P/Invoke signatures and memory management
- Improve test coverage with proper flow conservation checks
- Update documentation with detailed build instructions
- Configure NuGet package for Linux runtime support

+++ src/LemonNet/EdgeFlow.cs
@@ -10,9 +10,9 @@ public readonly struct EdgeFlow : IEquatable<EdgeFlow>
 {
     public Node Source { get; }
     public Node Target { get; }
-    public double Flow { get; }
+    public long Flow { get; }
 
-    public EdgeFlow(Node source, Node target, double flow)
+    public EdgeFlow(Node source, Node target, long flow)
     {
         Source = source;
         Target = target;
